chore: upgrade for anycable-rails 1.3.0
