+++ spec/graphql/anycable_spec.rb
@@ -24,7 +24,7 @@ RSpec.describe GraphQL::AnyCable do
 
   let(:channel) do
     socket = double("Socket", istate: AnyCable::Socket::State.new({}))
-    connection = double("Connection", socket: socket)
+    connection = double("Connection", anycable_socket: socket)
     double("Channel", id: "legacy_id", params: { "channelId" => "legacy_id" }, stream_from: nil, connection: connection)
   end
