+++ spec/integration_helper.rb
@@ -36,6 +36,8 @@ class FakeConnection
   attr_reader :request, :socket, :identifiers, :subscriptions,
               :schema
 
+  alias anycable_socket socket
+
   def initialize(socket, identifiers: nil, subscriptions: nil)
     @socket = socket
     @identifiers = identifiers ? JSON.parse(identifiers) : {}
